session: It is active only after first state now. More player properties are synced.

+++ src/api.h
@@ -32,7 +32,7 @@ typedef struct api_state_player_t {
 	uint32_t level;
 	uint32_t exp;
 	uint32_t money;
-	uint8_t  x, y;
+	uint32_t x, y;
 
 	api_state_resource_t mind;
 	api_state_resource_t matter;
